# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "domain.Notification": {
                "properties": {
                    "activityId": {
                        "type": "string"
                    },
                    "actorId": {
                        "type": "string"
                    },
                    "actorName": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "isRead": {
                        "type": "boolean"
                    },
                    "title": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "http.ListNotificationsResponse": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/domain.Notification"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/pagination.Meta"
                    }
                },
                "type": "object"
            },
...
            "pagination.Meta": {
                "properties": {
                    "hasMore": {
                        "type": "boolean"
                    },
                    "nextCursor": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "response.ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "response.MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "Backend API service for Splittr bill splitting.",
        "title": "Splittr API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/notifications": {
            "get": {
                "description": "Get a cursor-paginated list of notifications for the current user.",
                "parameters": [
                    {
                        "description": "Items per page (max 100, default 20)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque cursor token from a previous response",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ListNotificationsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List notifications",
                "tags": [
                    "notifications"
                ]
            }
        },
        "/notifications/read-all": {
            "post": {
                "description": "Mark all unread notifications as read for the current user.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.MessageResponse"
                                }
                            }
                        },
                        "description": "Success message"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Mark all notifications as read",
                "tags": [
                    "notifications"
                ]
            }
        },
        "/notifications/{id}/read": {
            "post": {
                "description": "Mark a specific notification as read by ID.",
                "parameters": [
                    {
                        "description": "Notification ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.MessageResponse"
                                }
                            }
                        },
                        "description": "Success message"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Mark notification as read",
                "tags": [
                    "notifications"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/v1"
        }
    ]
}